Enhance key management and UI functionality
- Introduced new modules for profile management, key handling, and UI rendering to improve code organization and maintainability.
- Updated `app.js` to integrate these new modules, enhancing the keybind management experience with improved view toggling and dynamic UI updates.
- Refactored CSS styles in `styles.css` to support new UI elements and ensure consistent layout across different views.
- Added comprehensive tests to validate the new functionality and ensure UI elements reflect changes accurately.

These updates significantly enhance the application's key management capabilities and overall user interface responsiveness, contributing to a more intuitive user experience.
