+++ src/js/app.js
@@ -4,6 +4,9 @@ import store from './store.js'
 import eventBus from './eventBus.js'
 import STOPreferencesManager from './preferences.js'
 import STOAutoSyncManager from './autoSync.js'
+import { profileManagement } from './profileManagement.js'
+import { keyHandling } from './keyHandling.js'
+import { uiRendering } from './uiRendering.js'
 
 export default class STOToolsKeybindManager {
   constructor() {
@@ -119,6 +122,10 @@ export default class STOToolsKeybindManager {
       // Update mode buttons to reflect current environment
       this.updateModeButtons()
 
+      // Update toggle button to reflect current view mode
+      const currentViewMode = localStorage.getItem('keyViewMode') || 'key-types'
+      this.updateViewToggleButton(currentViewMode)
+
       // Update theme toggle button to reflect current theme
       const settings = stoStorage.getSettings()
       this.updateThemeToggleButton(settings.theme || 'default')
@@ -148,1136 +155,6 @@ export default class STOToolsKeybindManager {
       eventBus.emit('sto-app-error', { error })
     }
   }
-
-  // Data Management
-  async loadData() {
-    const data = stoStorage.getAllData()
-    this.currentProfile = data.currentProfile
-
-    // Load current environment from profile or default to space
-    const profileData = data.profiles[this.currentProfile]
-    if (profileData) {
-      this.currentEnvironment = profileData.currentEnvironment || 'space'
-    } else {
-      this.currentEnvironment = 'space'
-    }
-
-    // Validate current profile exists
-    if (!data.profiles[this.currentProfile]) {
-      this.currentProfile = Object.keys(data.profiles)[0]
-      this.saveCurrentProfile()
-    }
-  }
-
-  saveData() {
-    const data = stoStorage.getAllData()
-    data.currentProfile = this.currentProfile
-    data.lastModified = new Date().toISOString()
-
-    if (stoStorage.saveAllData(data)) {
-      this.setModified(false)
-      return true
-    }
-    return false
-  }
-
-  saveCurrentProfile() {
-    const data = stoStorage.getAllData()
-    data.currentProfile = this.currentProfile
-    return stoStorage.saveAllData(data)
-  }
-
-  setModified(modified = true) {
-    this.isModified = modified
-    const indicator = document.getElementById('modifiedIndicator')
-    if (indicator) {
-      indicator.style.display = modified ? 'inline' : 'none'
-    }
-    
-    // Emit modification event for auto-sync
-    if (modified) {
-      eventBus.emit('profile-modified')
-    }
-  }
-
-  // Profile Management
-  getCurrentProfile() {
-    const profile = stoStorage.getProfile(this.currentProfile)
-    if (!profile) return null
-
-    // Return current environment build data
-    return this.getCurrentBuild(profile)
-  }
-
-  getCurrentBuild(profile) {
-    if (!profile) return null
-
-    // Ensure builds exist
-    if (!profile.builds) {
-      profile.builds = {
-        space: { keys: {} },
-        ground: { keys: {} },
-      }
-    }
-
-    if (!profile.builds[this.currentEnvironment]) {
-      profile.builds[this.currentEnvironment] = { keys: {} }
-    }
-
-    // Ensure the build keys object exists
-    if (!profile.builds[this.currentEnvironment].keys) {
-      profile.builds[this.currentEnvironment].keys = {}
-    }
-
-    // Return a profile-like object with current build data
-    // Keys are build-specific, but aliases are profile-wide
-    // IMPORTANT: keys must be a direct reference, not a copy
-    return {
-      ...profile,
-      keys: profile.builds[this.currentEnvironment].keys, // Direct reference to build keys
-      aliases: profile.aliases || {}, // Profile-level aliases, not build-specific
-    }
-  }
-
-  switchProfile(profileId) {
-    if (profileId !== this.currentProfile) {
-      this.currentProfile = profileId
-      this.selectedKey = null
-
-      const profile = stoStorage.getProfile(profileId)
-      if (profile) {
-        this.currentEnvironment = profile.currentEnvironment || 'space'
-      }
-
-      this.saveCurrentProfile()
-      this.renderKeyGrid()
-      this.renderCommandChain()
-      this.updateProfileInfo()
-
-      const currentBuild = this.getCurrentProfile()
-      stoUI.showToast(
-        i18next.t('switched_to_profile', { name: currentBuild.name, environment: this.currentEnvironment }),
-        'success'
-      )
-    }
-  }
-
-  createProfile(name, description = '', mode = 'space') {
-    const profileId = this.generateProfileId(name)
-    const profile = {
-      name,
-      description,
-      currentEnvironment: mode,
-      builds: {
-        space: { keys: {} },
-        ground: { keys: {} },
-      },
-      aliases: {},
-      created: new Date().toISOString(),
-      lastModified: new Date().toISOString(),
-    }
-
-    if (stoStorage.saveProfile(profileId, profile)) {
-      this.switchProfile(profileId)
-      this.renderProfiles()
-      stoUI.showToast(i18next.t('profile_created', {name: name}), 'success')
-      return profileId
-    }
-
-    stoUI.showToast(i18next.t('failed_to_create_profile'), 'error')
-    return null
-  }
-
-  cloneProfile(sourceProfileId, newName) {
-    const sourceProfile = stoStorage.getProfile(sourceProfileId)
-    if (!sourceProfile) return null
-
-    const profileId = this.generateProfileId(newName)
-    const clonedProfile = {
-      ...JSON.parse(JSON.stringify(sourceProfile)), // Deep clone
-      name: newName,
-      description: `Copy of ${sourceProfile.name}`,
-      created: new Date().toISOString(),
-      lastModified: new Date().toISOString(),
-    }
-
-    if (stoStorage.saveProfile(profileId, clonedProfile)) {
-      this.renderProfiles()
-      stoUI.showToast(
-        i18next.t('profile_created_from', { newName: newName, sourceProfile: sourceProfile.name }),
-        'success'
-      )
-      return profileId
-    }
-
-    stoUI.showToast(i18next.t('failed_to_clone_profile'), 'error')
-    return null
-  }
-
-  deleteProfile(profileId) {
-    const profile = stoStorage.getProfile(profileId)
-    if (!profile) return false
-
-    const data = stoStorage.getAllData()
-    const profileCount = Object.keys(data.profiles).length
-
-    if (profileCount <= 1) {
-      stoUI.showToast(i18next.t('cannot_delete_the_last_profile'), 'warning')
-      return false
-    }
-
-    if (stoStorage.deleteProfile(profileId)) {
-      if (this.currentProfile === profileId) {
-        // Switch to first available profile
-        const remainingProfiles = Object.keys(stoStorage.getAllData().profiles)
-        this.switchProfile(remainingProfiles[0])
-      }
-
-      this.renderProfiles()
-      stoUI.showToast(i18next.t('profile_deleted', {profileName: profile.name}), 'success')
-      return true
-    }
-
-    stoUI.showToast(i18next.t('failed_to_delete_profile'), 'error')
-    return false
-  }
-
-  generateProfileId(name) {
-    const base = name.toLowerCase().replace(/[^a-z0-9]/g, '_')
-    let id = base
-    let counter = 1
-
-    const data = stoStorage.getAllData()
-    while (data.profiles[id]) {
-      id = `${base}_${counter}`
-      counter++
-    }
-
-    return id
-  }
-
-  // Key Management
-  selectKey(keyName) {
-    this.selectedKey = keyName
-
-    // Update stabilize checkbox based on keybind metadata (environment-scoped)
-    const profile = this.getCurrentProfile()
-    const stabilizeCheckbox = document.getElementById('stabilizeExecutionOrder')
-    if (stabilizeCheckbox && profile && profile.keybindMetadata) {
-      let flag = false
-      if (profile.keybindMetadata[this.currentEnvironment]) {
-        const envMeta = profile.keybindMetadata[this.currentEnvironment]
-        flag = !!(envMeta[keyName] && envMeta[keyName].stabilizeExecutionOrder)
-      }
-      stabilizeCheckbox.checked = flag
-    } else if (stabilizeCheckbox) {
-      stabilizeCheckbox.checked = false
-    }
-
-    this.renderKeyGrid()
-    this.renderCommandChain()
-    this.updateChainActions()
-  }
-
-  addKey(keyName) {
-    if (!this.isValidKeyName(keyName)) {
-      stoUI.showToast(i18next.t('invalid_key_name'), 'error')
-      return false
-    }
-
-    // Get the full profile from storage
-    const fullProfile = stoStorage.getProfile(this.currentProfile)
-    if (!fullProfile) {
-      stoUI.showToast(i18next.t('no_profile_selected'), 'error')
-      return false
-    }
-
-    // Ensure builds structure exists
-    if (!fullProfile.builds) {
-      fullProfile.builds = {
-        space: { keys: {} },
-        ground: { keys: {} },
-      }
-    }
-
-    // Ensure current environment build exists
-    if (!fullProfile.builds[this.currentEnvironment]) {
-      fullProfile.builds[this.currentEnvironment] = { keys: {} }
-    }
-
-    // Check if key already exists in the current build
-    if (!fullProfile.builds[this.currentEnvironment].keys[keyName]) {
-      // Add the key to the current build
-      fullProfile.builds[this.currentEnvironment].keys[keyName] = []
-      
-      // Save the full profile back to storage
-      stoStorage.saveProfile(this.currentProfile, fullProfile)
-      
-      // Update UI
-      this.renderKeyGrid()
-      this.selectKey(keyName)
-      this.setModified(true)
-
-      stoUI.showToast(i18next.t('key_added', {keyName: keyName}), 'success')
-      return true
-    } else {
-      stoUI.showToast(i18next.t('key_already_exists', {keyName: keyName}), 'warning')
-      return false
-    }
-  }
-
-  deleteKey(keyName) {
-    // Get the full profile from storage
-    const fullProfile = stoStorage.getProfile(this.currentProfile)
-    if (!fullProfile) {
-      stoUI.showToast(i18next.t('no_profile_selected'), 'error')
-      return false
-    }
-
-    // Check if key exists in the current build
-    if (fullProfile.builds && 
-        fullProfile.builds[this.currentEnvironment] && 
-        fullProfile.builds[this.currentEnvironment].keys && 
-        fullProfile.builds[this.currentEnvironment].keys[keyName]) {
-      
-      // Delete the key from the current build
-      delete fullProfile.builds[this.currentEnvironment].keys[keyName]
-      
-      // Save the full profile back to storage
-      stoStorage.saveProfile(this.currentProfile, fullProfile)
-
-      if (this.selectedKey === keyName) {
-        this.selectedKey = null
-      }
-
-      this.renderKeyGrid()
-      this.renderCommandChain()
-      this.setModified(true)
-
-      stoUI.showToast(i18next.t('key_deleted', {keyName: keyName}), 'success')
-      return true
-    }
-
-    return false
-  }
-
-  isValidKeyName(keyName) {
-    return (
-      STO_DATA.validation.keyNamePattern.test(keyName) && keyName.length <= 20
-    )
-  }
-
-  // Command Management
-  addCommand(keyName, command) {
-    if (this.currentEnvironment === 'alias') {
-      // Handle alias command addition - aliases are profile-level
-      const profile = this.getCurrentProfile()
-      
-      if (!profile.aliases) {
-        profile.aliases = {}
-      }
-      if (!profile.aliases[keyName]) {
-        profile.aliases[keyName] = { description: '', commands: '' }
-      }
-
-      // Add command to alias
-      const currentCommands = profile.aliases[keyName].commands
-      const newCommand = Array.isArray(command) ? 
-        command.map(cmd => cmd.command).join(' $$ ') : 
-        command.command
-
-      if (currentCommands) {
-        profile.aliases[keyName].commands = currentCommands + ' $$ ' + newCommand
-      } else {
-        profile.aliases[keyName].commands = newCommand
-      }
-
-      stoStorage.saveProfile(this.currentProfile, profile)
-      stoUI.showToast(i18next.t('command_added_to_alias'), 'success')
-    } else {
-      // Handle keybind command addition - keybinds are build-specific
-      const fullProfile = stoStorage.getProfile(this.currentProfile)
-      if (!fullProfile) {
-        stoUI.showToast(i18next.t('no_profile_selected'), 'error')
-        return
-      }
-
-      // Ensure builds structure exists
-      if (!fullProfile.builds) {
-        fullProfile.builds = {
-          space: { keys: {} },
-          ground: { keys: {} },
-        }
-      }
-
-      // Ensure current environment build exists
-      if (!fullProfile.builds[this.currentEnvironment]) {
-        fullProfile.builds[this.currentEnvironment] = { keys: {} }
-      }
-
-      // Ensure the key exists
-      if (!fullProfile.builds[this.currentEnvironment].keys[keyName]) {
-        fullProfile.builds[this.currentEnvironment].keys[keyName] = []
-      }
-
-      // Handle both single commands and arrays of commands (for tray ranges)
-      if (Array.isArray(command)) {
-        // Array of commands - add each one with unique ID if not already present
-        command.forEach((cmd) => {
-          if (!cmd.id) {
-            cmd.id = this.generateCommandId()
-          }
-          fullProfile.builds[this.currentEnvironment].keys[keyName].push(cmd)
-        })
-        stoUI.showToast(i18next.t('commands_added', {count: command.length}), 'success')
-      } else {
-        // Single command
-        if (!command.id) {
-          command.id = this.generateCommandId()
-        }
-        fullProfile.builds[this.currentEnvironment].keys[keyName].push(command)
-        stoUI.showToast(i18next.t('command_added'), 'success')
-      }
-
-      // Save the full profile back to storage
-      stoStorage.saveProfile(this.currentProfile, fullProfile)
-    }
-
-    this.renderCommandChain()
-    if (this.currentEnvironment === 'alias') {
-      this.renderAliasGrid()
-    } else {
-      this.renderKeyGrid()
-    }
-    this.setModified(true)
-    
-    // Emit command modification event for auto-sync
-    eventBus.emit('command-modified', { keyName, command, action: 'add' })
-  }
-
-  deleteCommand(keyName, commandIndex) {
-    if (this.currentEnvironment === 'alias') {
-      // Handle alias command deletion - aliases are profile-level
-      const profile = this.getCurrentProfile()
-      
-      const alias = profile.aliases && profile.aliases[keyName]
-      if (alias && alias.commands) {
-        const commands = alias.commands.split('$$').map(cmd => cmd.trim())
-        if (commandIndex >= 0 && commandIndex < commands.length) {
-          const deletedCommand = commands[commandIndex]
-          commands.splice(commandIndex, 1)
-          profile.aliases[keyName].commands = commands.join(' $$ ')
-          
-          stoStorage.saveProfile(this.currentProfile, profile)
-          this.renderCommandChain()
-          this.renderAliasGrid()
-          this.setModified(true)
-
-          stoUI.showToast(i18next.t('command_deleted_from_alias'), 'success')
-          
-          // Emit command modification event for auto-sync
-          eventBus.emit('command-modified', { keyName, command: { command: deletedCommand }, action: 'delete' })
-        }
-      }
-    } else {
-      // Handle keybind command deletion - keybinds are build-specific
-      const fullProfile = stoStorage.getProfile(this.currentProfile)
-      if (!fullProfile) {
-        stoUI.showToast(i18next.t('no_profile_selected'), 'error')
-        return
-      }
-
-      if (fullProfile.builds && 
-          fullProfile.builds[this.currentEnvironment] && 
-          fullProfile.builds[this.currentEnvironment].keys && 
-          fullProfile.builds[this.currentEnvironment].keys[keyName] && 
-          fullProfile.builds[this.currentEnvironment].keys[keyName][commandIndex]) {
-        
-        const deletedCommand = fullProfile.builds[this.currentEnvironment].keys[keyName][commandIndex]
-        fullProfile.builds[this.currentEnvironment].keys[keyName].splice(commandIndex, 1)
-        
-        // Save the full profile back to storage
-        stoStorage.saveProfile(this.currentProfile, fullProfile)
-        this.renderCommandChain()
-        this.renderKeyGrid()
-        this.setModified(true)
-
-        stoUI.showToast(i18next.t('command_deleted'), 'success')
-        
-        // Emit command modification event for auto-sync
-        eventBus.emit('command-modified', { keyName, command: deletedCommand, action: 'delete' })
-      }
-    }
-  }
-
-  moveCommand(keyName, fromIndex, toIndex) {
-    if (this.currentEnvironment === 'alias') {
-      // Handle alias command movement - aliases are profile-level
-      const profile = this.getCurrentProfile()
-      
-      const alias = profile.aliases && profile.aliases[keyName]
-      if (alias && alias.commands) {
-        const commands = alias.commands.split('$$').map(cmd => cmd.trim())
-        
-        if (
-          fromIndex >= 0 &&
-          fromIndex < commands.length &&
-          toIndex >= 0 &&
-          toIndex < commands.length
-        ) {
-          const [command] = commands.splice(fromIndex, 1)
-          commands.splice(toIndex, 0, command)
-          
-          profile.aliases[keyName].commands = commands.join(' $$ ')
-          stoStorage.saveProfile(this.currentProfile, profile)
-          this.renderCommandChain()
-          this.setModified(true)
-          
-          // Emit command modification event for auto-sync
-          eventBus.emit('command-modified', { keyName, command: { command }, action: 'move', fromIndex, toIndex })
-        }
-      }
-    } else {
-      // Handle keybind command movement - keybinds are build-specific
-      const fullProfile = stoStorage.getProfile(this.currentProfile)
-      if (!fullProfile) {
-        stoUI.showToast(i18next.t('no_profile_selected'), 'error')
-        return
-      }
-
-      const commands = fullProfile.builds && 
-                      fullProfile.builds[this.currentEnvironment] && 
-                      fullProfile.builds[this.currentEnvironment].keys && 
-                      fullProfile.builds[this.currentEnvironment].keys[keyName]
-
-      if (
-        commands &&
-        fromIndex >= 0 &&
-        fromIndex < commands.length &&
-        toIndex >= 0 &&
-        toIndex < commands.length
-      ) {
-        const [command] = commands.splice(fromIndex, 1)
-        commands.splice(toIndex, 0, command)
-
-        // Save the full profile back to storage
-        stoStorage.saveProfile(this.currentProfile, fullProfile)
-        this.renderCommandChain()
-        this.setModified(true)
-        
-        // Emit command modification event for auto-sync
-        eventBus.emit('command-modified', { keyName, command, action: 'move', fromIndex, toIndex })
-      }
-    }
-  }
-
-  generateCommandId() {
-    return 'cmd_' + Date.now() + '_' + Math.random().toString(36).substr(2, 9)
-  }
-
-  // Rendering Methods
-  renderProfiles() {
-    const select = document.getElementById('profileSelect')
-    if (!select) return
-
-    const data = stoStorage.getAllData()
-    select.innerHTML = ''
-
-    if (Object.keys(data.profiles).length === 0) {
-      // No profiles available
-      const option = document.createElement('option')
-      option.value = ''
-      option.textContent = 'No profiles available'
-      option.disabled = true
-      select.appendChild(option)
-    } else {
-      Object.entries(data.profiles).forEach(([id, profile]) => {
-        const option = document.createElement('option')
-        option.value = id
-        option.textContent = profile.name
-        if (id === this.currentProfile) {
-          option.selected = true
-        }
-        select.appendChild(option)
-      })
-    }
-
-    this.updateProfileInfo()
-  }
-
-  updateProfileInfo() {
-    const profile = this.getCurrentProfile()
-
-    // Update mode buttons - only update active state, don't disable them
-    const modeBtns = document.querySelectorAll('.mode-btn')
-    modeBtns.forEach((btn) => {
-      btn.classList.toggle(
-        'active',
-        profile && btn.dataset.mode === this.currentEnvironment
-      )
-      // Remove the problematic btn.disabled = !profile line
-      // Mode buttons should always be enabled as long as we have a valid profile ID
-      btn.disabled = !this.currentProfile
-    })
-
-    // Update key count
-    const keyCount = document.getElementById('keyCount')
-    if (keyCount) {
-      if (profile) {
-        const count = Object.keys(profile.keys).length
-        const keyText = count === 1 ? i18next.t('key') : i18next.t('keys')
-        keyCount.textContent = `${count} ${keyText}`
-      } else {
-        keyCount.textContent = i18next.t('no_profile')
-      }
-    }
-  }
-
-  renderKeyGrid() {
-    const grid = document.getElementById('keyGrid')
-    if (!grid) return
-
-    const profile = this.getCurrentProfile()
-    if (!profile) return
-
-    grid.innerHTML = ''
-
-    if (!profile) {
-      // Show empty state when no profile is available
-      grid.innerHTML = `
-                <div class="empty-state">
-                    <i class="fas fa-folder-open"></i>
-                    <h4>No Profile Selected</h4>
-                    <p>Create a new profile or load default data to get started.</p>
-                </div>
-            `
-      return
-    }
-
-    // Get all keys for this profile
-    const keys = Object.keys(profile.keys)
-    // console.log('renderKeyGrid: Profile keys:', keys);
-    // console.log('renderKeyGrid: Profile keys count:', keys.length);
-
-    // Add common keys that might not have commands yet
-    const commonKeys = [
-      'Space',
-      '1',
-      '2',
-      '3',
-      '4',
-      '5',
-      'F1',
-      'F2',
-      'F3',
-      'F4',
-    ]
-    const allKeys = [...new Set([...keys, ...commonKeys])].sort()
-    // console.log('renderKeyGrid: All keys to render:', allKeys);
-
-    // Check view preference
-    const viewMode = localStorage.getItem('keyViewMode') || 'key-types'
-
-    if (viewMode === 'categorized') {
-      this.renderCategorizedKeyView(grid, profile, allKeys)
-    } else if (viewMode === 'key-types') {
-      this.renderKeyTypeView(grid, profile, allKeys)
-    } else {
-      this.renderSimpleKeyGrid(grid, allKeys)
-    }
-
-    // Update toggle button icon
-    this.updateViewToggleButton(viewMode)
-
-    // Reapply any existing filter after rendering the new view
-    const filterInput = document.getElementById('keyFilter')
-    if (filterInput && filterInput.value.trim()) {
-      this.filterKeys(filterInput.value.trim())
-    }
-  }
-
-  renderSimpleKeyGrid(grid, allKeys) {
-    // Remove categorized class to use normal grid layout
-    grid.classList.remove('categorized')
-
-    allKeys.forEach((keyName) => {
-      const keyElement = this.createKeyElement(keyName)
-      grid.appendChild(keyElement)
-    })
-  }
-
-  renderCategorizedKeyView(grid, profile, allKeys) {
-    // Add categorized class to override grid layout
-    grid.classList.add('categorized')
-
-    // Categorize keys based on their commands, including all keys
-    const categorizedKeys = this.categorizeKeys(profile.keys, allKeys)
-
-    // Sort categories by priority (Unknown first, then alphabetically)
-    const sortedCategories = Object.entries(categorizedKeys).sort(
-      ([aId, aData], [bId, bData]) => {
-        if (aData.priority !== bData.priority) {
-          return aData.priority - bData.priority
-        }
-        return aData.name.localeCompare(bData.name)
-      }
-    )
-
-    // Create category tree structure - show all categories even if empty
-    sortedCategories.forEach(([categoryId, categoryData]) => {
-      const categoryElement = this.createKeyCategoryElement(
-        categoryId,
-        categoryData
-      )
-      grid.appendChild(categoryElement)
-    })
-  }
-
-  categorizeKeys(keysWithCommands, allKeys) {
-    const categories = {}
-
-    // Add Unknown category first (for empty keys)
-    categories.unknown = {
-      name: 'Unknown',
-      icon: 'fas fa-question-circle',
-      keys: new Set(),
-      priority: 0,
-    }
-
-    // Initialize categories from command library
-    Object.entries(STO_DATA.commands).forEach(([categoryId, categoryData]) => {
-      categories[categoryId] = {
-        name: categoryData.name,
-        icon: categoryData.icon,
-        keys: new Set(),
-        priority: 1,
-      }
-    })
-
-    // Process all keys, not just ones with commands
-    allKeys.forEach((keyName) => {
-      const commands = keysWithCommands[keyName] || []
-
-      if (!commands || commands.length === 0) {
-        categories.unknown.keys.add(keyName)
-        return
-      }
-
-      // Get all categories this key belongs to
-      const keyCategories = new Set()
-      commands.forEach((command) => {
-        if (command.type && categories[command.type]) {
-          keyCategories.add(command.type)
-        } else if (window.stoCommands) {
-          // Use command detection if type is not set
-          const detectedType = window.stoCommands.detectCommandType(
-            command.command
-          )
-          if (categories[detectedType]) {
-            keyCategories.add(detectedType)
-          }
-        }
-      })
-
-      // Add key to all relevant categories
-      if (keyCategories.size > 0) {
-        keyCategories.forEach((categoryId) => {
-          categories[categoryId].keys.add(keyName)
-        })
-      } else {
-        // Fallback for unknown command types
-        if (!categories.custom) {
-          categories.custom = {
-            name: 'Custom Commands',
-            icon: 'fas fa-cog',
-            keys: new Set(),
-            priority: 2,
-          }
-        }
-        categories.custom.keys.add(keyName)
-      }
-    })
-
-    // Convert sets to arrays and sort
-    Object.values(categories).forEach((category) => {
-      category.keys = Array.from(category.keys).sort(
-        this.compareKeys.bind(this)
-      )
-    })
-
-    return categories
-  }
-
-  createKeyCategoryElement(categoryId, categoryData, mode = 'command') {
-    const element = document.createElement('div')
-    element.className = 'category'
-    element.dataset.category = categoryId
-
-    // Use different storage key for key-type categories
-    const storageKey =
-      mode === 'key-type'
-        ? `keyTypeCategory_${categoryId}_collapsed`
-        : `keyCategory_${categoryId}_collapsed`
-    const isCollapsed = localStorage.getItem(storageKey) === 'true'
-
-    element.innerHTML = `
-            <h4 class="${isCollapsed ? 'collapsed' : ''}" data-category="${categoryId}" data-mode="${mode}">
-                <i class="fas fa-chevron-right category-chevron"></i>
-                <i class="${categoryData.icon}"></i> 
-                ${categoryData.name} 
-                <span class="key-count">(${categoryData.keys.length})</span>
-            </h4>
-            <div class="category-commands ${isCollapsed ? 'collapsed' : ''}">
-                ${categoryData.keys.map((keyName) => this.createKeyElementHTML(keyName)).join('')}
-            </div>
-        `
-
-    // Add click handler for category header
-    const header = element.querySelector('h4')
-    header.addEventListener('click', () => {
-      this.toggleKeyCategory(categoryId, element, mode)
-    })
-
-    // Add click handlers for key elements
-    const keyElements = element.querySelectorAll('.command-item')
-    keyElements.forEach((keyElement) => {
-      keyElement.addEventListener('click', () => {
-        const keyName = keyElement.dataset.key
-        this.selectKey(keyName)
-      })
-    })
-
-    return element
-  }
-
-  createKeyElementHTML(keyName) {
-    const profile = this.getCurrentProfile()
-    const commands = profile.keys[keyName] || []
-    const isActive = this.selectedKey === keyName
-
-    // Simple length-based font sizing for categorized view (no line breaks)
-    const keyLength = keyName.length
-    let lengthClass
-    if (keyLength <= 6) {
-      lengthClass = 'short'
-    } else if (keyLength <= 12) {
-      lengthClass = 'medium'
-    } else {
-      lengthClass = 'long'
-    }
-
-    return `
-            <div class="command-item ${isActive ? 'active' : ''}" data-key="${keyName}" data-length="${lengthClass}">
-                <span class="key-label">${keyName}</span>
-                ${
-                  commands.length > 0
-                    ? `
-                    <span class="command-count-badge">${commands.length}</span>
-                `
-                    : ''
-                }
-            </div>
-        `
-  }
-
-  toggleKeyCategory(categoryId, element, mode = 'command') {
-    const header = element.querySelector('h4')
-    const commands = element.querySelector('.category-commands')
-    const chevron = header.querySelector('.category-chevron')
-
-    const isCollapsed = commands.classList.contains('collapsed')
-
-    // Use different storage key for key-type categories
-    const storageKey =
-      mode === 'key-type'
-        ? `keyTypeCategory_${categoryId}_collapsed`
-        : `keyCategory_${categoryId}_collapsed`
-
-    if (isCollapsed) {
-      commands.classList.remove('collapsed')
-      header.classList.remove('collapsed')
-      chevron.style.transform = 'rotate(90deg)'
-      localStorage.setItem(storageKey, 'false')
-    } else {
-      commands.classList.add('collapsed')
-      header.classList.add('collapsed')
-      chevron.style.transform = 'rotate(0deg)'
-      localStorage.setItem(storageKey, 'true')
-    }
-  }
-
-  formatKeyName(keyName) {
-    // Smart formatting for compound keys to use line breaks instead of tiny text
-    if (keyName.includes('+')) {
-      // Always break compound keys for better readability
-      return keyName.replace(/\+/g, '<br>+<br>')
-    }
-    // Handle underscores for word wrapping
-    if (keyName.includes('_')) {
-      return keyName.replace(/_/g, '<br>_<br>')
-    }
-    return keyName
-  }
-
-  detectKeyTypes(keyName) {
-    const types = []
-
-    // Check if this is a compound key with modifiers
-    if (keyName.includes('+')) {
-      const parts = keyName.split('+')
-      const hasModifier = parts.some((part) =>
-        part.match(/^(Ctrl|Control|Alt|Shift|Win|Cmd|Super)$/i)
-      )
-
-      if (hasModifier) {
-        types.push('modifiers')
-      }
-    }
-
-    // Extract the base key from compound keys to check its type too
-    const baseKey = keyName.split('+').pop()
-
-    // Function Keys
-    if (baseKey.match(/^F\d+$/)) types.push('function')
-    // Mouse inputs - comprehensive list from STO documentation
-    else if (
-      baseKey.match(
-        /^(Lbutton|Rbutton|Mbutton|Leftdrag|Rightdrag|Middledrag|Leftclick|Rightclick|Middleclick|Leftdoubleclick|Rightdoubleclick|Middledoubleclick|Wheelplus|Wheelminus|Mousechord|Mouse|Wheel|LMouse|RMouse|MMouse|XMouse|Drag)/i
-      )
-    )
-      types.push('mouse')
-    // Numberpad
-    else if (baseKey.match(/^(Numpad|Keypad)/i)) types.push('numberpad')
-    // Single Modifiers (for standalone modifier keys)
-    else if (baseKey.match(/^(Ctrl|Control|Alt|Shift|Win|Cmd|Super)$/i)) {
-      if (!types.includes('modifiers')) types.push('modifiers')
-    }
-
-    // Navigation
-    else if (
-      baseKey.match(
-        /^(Up|Down|Left|Right|Home|End|PageUp|PageDown|Insert|Delete)$/i
-      )
-    )
-      types.push('navigation')
-    // System/Special
-    else if (
-      baseKey.match(
-        /^(Space|Tab|Enter|Return|Escape|Esc|Backspace|CapsLock|ScrollLock|NumLock|PrintScreen|Pause|Break)$/i
-      )
-    )
-      types.push('system')
-    // Numbers
-    else if (baseKey.match(/^[0-9]$/)) types.push('alphanumeric')
-    // Letters
-    else if (baseKey.match(/^[A-Za-z]$/)) types.push('alphanumeric')
-    // Symbols and punctuation
-    else if (
-      baseKey.match(/^[`~!@#$%^&*()_+\-=\[\]{}\\|;':",./<>?]$/) ||
-      baseKey.match(
-        /^(Comma|Period|Semicolon|Quote|Slash|Backslash|Minus|Plus|Equals|Bracket|Grave|Tilde)$/i
-      )
-    ) {
-      types.push('symbols')
-    }
-
-    // Default fallback
-    else {
-      types.push('other')
-    }
-
-    return types.length > 0 ? types : ['other']
-  }
-
-  categorizeKeysByType(keysWithCommands, allKeys) {
-    const categories = {
-      function: {
-        name: 'Function Keys',
-        icon: 'fas fa-keyboard',
-        keys: new Set(),
-        priority: 1,
-      },
-      alphanumeric: {
-        name: 'Letters & Numbers',
-        icon: 'fas fa-font',
-        keys: new Set(),
-        priority: 2,
-      },
-      numberpad: {
-        name: 'Numberpad',
-        icon: 'fas fa-calculator',
-        keys: new Set(),
-        priority: 3,
-      },
-      modifiers: {
-        name: 'Modifier Keys',
-        icon: 'fas fa-hand-paper',
-        keys: new Set(),
-        priority: 4,
-      },
-      navigation: {
-        name: 'Navigation',
-        icon: 'fas fa-arrows-alt',
-        keys: new Set(),
-        priority: 5,
-      },
-      system: {
-        name: 'System Keys',
-        icon: 'fas fa-cogs',
-        keys: new Set(),
-        priority: 6,
-      },
-      mouse: {
-        name: 'Mouse & Wheel',
-        icon: 'fas fa-mouse',
-        keys: new Set(),
-        priority: 7,
-      },
-      symbols: {
-        name: 'Symbols & Punctuation',
-        icon: 'fas fa-at',
-        keys: new Set(),
-        priority: 8,
-      },
-      other: {
-        name: 'Other Keys',
-        icon: 'fas fa-question-circle',
-        keys: new Set(),
-        priority: 9,
-      },
-    }
-
-    // Categorize each key by its types (can be multiple)
-    allKeys.forEach((keyName) => {
-      const keyTypes = this.detectKeyTypes(keyName)
-      keyTypes.forEach((keyType) => {
-        if (categories[keyType]) {
-          categories[keyType].keys.add(keyName)
-        } else {
-          categories.other.keys.add(keyName)
-        }
-      })
-    })
-
-    // Convert sets to arrays and sort
-    Object.values(categories).forEach((category) => {
-      category.keys = Array.from(category.keys).sort(
-        this.compareKeys.bind(this)
-      )
-    })
-
-    return categories
-  }
-
-  renderKeyTypeView(grid, profile, allKeys) {
-    // Add categorized class to override grid layout
-    grid.classList.add('categorized')
-
-    // Categorize keys by their input type
-    const categorizedKeys = this.categorizeKeysByType(profile.keys, allKeys)
-
-    // Sort categories by priority
-    const sortedCategories = Object.entries(categorizedKeys).sort(
-      ([aId, aData], [bId, bData]) => {
-        return aData.priority - bData.priority
-      }
-    )
-
-    // Create category tree structure - show all categories even if empty
-    sortedCategories.forEach(([categoryId, categoryData]) => {
-      const categoryElement = this.createKeyCategoryElement(
-        categoryId,
-        categoryData,
-        'key-type'
-      )
-      grid.appendChild(categoryElement)
-    })
-  }
-
-  compareKeys(a, b) {
-    // Custom key sorting logic
-    const getKeyPriority = (key) => {
-      if (key === 'Space') return 0
-      if (key.match(/^[0-9]$/)) return 1
-      if (key.match(/^F[0-9]+$/)) return 2
-      if (key.includes('Ctrl+')) return 3
-      if (key.includes('Alt+')) return 4
-      if (key.includes('Shift+')) return 5
-      return 6
-    }
-
-    const priorityA = getKeyPriority(a)
-    const priorityB = getKeyPriority(b)
-
-    if (priorityA !== priorityB) {
-      return priorityA - priorityB
-    }
-
-    return a.localeCompare(b)
-  }
-
-  createKeyElement(keyName) {
-    const profile = this.getCurrentProfile()
-    const commands = profile.keys[keyName] || []
-    const isSelected = keyName === this.selectedKey
-
-    const keyElement = document.createElement('div')
-    keyElement.className = `key-item ${isSelected ? 'active' : ''}`
-    keyElement.dataset.key = keyName
-    keyElement.title = `${keyName}: ${commands.length} command${commands.length !== 1 ? 's' : ''}`
-
-    // Smart formatting for compound keys and font sizing
-    const formattedKeyName = this.formatKeyName(keyName)
-    const hasLineBreaks = formattedKeyName.includes('<br>')
-
-    // Determine length classification
-    let lengthClass
-    if (hasLineBreaks) {
-      // For compound keys with line breaks, check the longest part
-      const parts = keyName.split(/[+_]/)
-      const longestPart = Math.max(...parts.map((part) => part.length))
-      if (longestPart <= 4) {
-        lengthClass = 'short'
-      } else if (longestPart <= 8) {
-        lengthClass = 'medium'
-      } else {
-        lengthClass = 'long'
-      }
-    } else {
-      // For single keys, use total length
-      const keyLength = keyName.length
-      if (keyLength <= 3) {
-        lengthClass = 'short'
-      } else if (keyLength <= 5) {
-        lengthClass = 'medium'
-      } else if (keyLength <= 8) {
-        lengthClass = 'long'
-      } else {
-        lengthClass = 'extra-long'
-      }
-    }
-
-    keyElement.dataset.length = lengthClass
-
-    keyElement.innerHTML = `
-            <div class="key-label">${formattedKeyName}</div>
-            ${
-              commands.length > 0
-                ? `
-                <div class="activity-bar" style="width: ${Math.min(commands.length * 15, 100)}%"></div>
-                <div class="command-count-badge">${commands.length}</div>
-            `
-                : ''
-            }
-        `
-
-    keyElement.addEventListener('click', () => {
-      this.selectKey(keyName)
-    })
-
-    return keyElement
-  }
-
   updateViewToggleButton(viewMode) {
     const toggleBtn = document.getElementById('toggleKeyViewBtn')
     if (toggleBtn) {
@@ -1296,6 +173,11 @@ export default class STOToolsKeybindManager {
   }
 
   toggleKeyView() {
+    // Only allow view mode changes in keybind mode, not alias mode
+    if (this.currentEnvironment === 'alias') {
+      return
+    }
+    
     const currentMode = localStorage.getItem('keyViewMode') || 'key-types'
     let newMode
 
@@ -1310,6 +192,7 @@ export default class STOToolsKeybindManager {
 
     localStorage.setItem('keyViewMode', newMode)
     this.renderKeyGrid()
+    this.updateViewToggleButton(newMode)
   }
 
   renderCommandChain() {
@@ -2204,278 +1087,16 @@ export default class STOToolsKeybindManager {
       this.updateChainOptionsForKeybind()
       this.filterCommandLibrary() // Apply environment filter to command library
     }
-  }
-
-  showAliasView() {
-    const keyContainer = document.querySelector('.key-selector-container')
-    const aliasContainer = document.getElementById('aliasSelectorContainer')
     
-    if (keyContainer) keyContainer.style.display = 'none'
-    if (aliasContainer) aliasContainer.style.display = 'block'
+    // Update toggle button visibility based on environment
+    this.updateToggleButtonVisibility()
   }
 
-  showKeyView() {
-    const keyContainer = document.querySelector('.key-selector-container')
-    const aliasContainer = document.getElementById('aliasSelectorContainer')
-    
-    if (keyContainer) keyContainer.style.display = 'block'
-    if (aliasContainer) aliasContainer.style.display = 'none'
-  }
-
-  updateChainOptionsForAlias() {
-    const stabilizeBtn = document.getElementById('stabilizeExecutionOrderBtn')
-    const aliasOptionsBtn = document.getElementById('aliasOptionsBtn')
-    const aliasOptions = document.getElementById('aliasOptions')
-    
-    if (stabilizeBtn) stabilizeBtn.style.display = 'none'
-    if (aliasOptionsBtn) aliasOptionsBtn.style.display = 'block'
-    if (aliasOptions) aliasOptions.style.display = 'none' // Initially hidden, shown when button is clicked
-  }
-
-  updateChainOptionsForKeybind() {
-    const stabilizeBtn = document.getElementById('stabilizeExecutionOrderBtn')
-    const aliasOptionsBtn = document.getElementById('aliasOptionsBtn')
-    const aliasOptions = document.getElementById('aliasOptions')
-    
-    if (stabilizeBtn) stabilizeBtn.style.display = 'block'
-    if (aliasOptionsBtn) aliasOptionsBtn.style.display = 'none'
-    if (aliasOptions) aliasOptions.style.display = 'none'
-  }
-
-  renderAliasGrid() {
-    const grid = document.getElementById('aliasGrid')
-    if (!grid) return
-
-    const profile = stoStorage.getProfile(this.currentProfile)
-    if (!profile || !profile.aliases) {
-      grid.innerHTML = `
-        <div class="empty-state">
-          <i class="fas fa-mask"></i>
-          <h4 data-i18n="no_aliases_defined">No aliases defined</h4>
-          <p data-i18n="create_alias_to_get_started">Create an alias to get started</p>
-        </div>
-      `
-      return
-    }
-
-    const aliases = Object.entries(profile.aliases)
-    if (aliases.length === 0) {
-      grid.innerHTML = `
-        <div class="empty-state">
-          <i class="fas fa-mask"></i>
-          <h4 data-i18n="no_aliases_defined">No aliases defined</h4>
-          <p data-i18n="create_alias_to_get_started">Create an alias to get started</p>
-        </div>
-      `
-      return
-    }
-
-    grid.innerHTML = aliases.map(([name, alias]) => 
-      this.createAliasChainElement(name, alias)
-    ).join('')
-
-    // Add event listeners to alias elements
-    grid.querySelectorAll('.alias-chain-item').forEach((item) => {
-      item.addEventListener('click', () => {
-        this.selectAlias(item.dataset.alias)
-      })
-    })
-  }
-
-  createAliasChainElement(name, alias) {
-    const commandCount = alias.commands ? alias.commands.split('$$').length : 0
-    const isSelected = this.selectedKey === name // Reuse selectedKey for alias selection
-    const description = alias.description || ''
-    
-    // Calculate length class for dynamic font sizing (similar to key elements)
-    // Since aliases don't use + separators like keys, use simple length-based logic
-    const nameLength = name.length
-    let lengthClass
-    if (nameLength <= 8) {
-      lengthClass = 'short'
-    } else if (nameLength <= 12) {
-      lengthClass = 'medium'
-    } else if (nameLength <= 16) {
-      lengthClass = 'long'
-    } else {
-      lengthClass = 'extra-long'
-    }
-    
-    return `
-      <div class="alias-chain-item ${isSelected ? 'selected' : ''}" data-alias="${name}" data-length="${lengthClass}" title="${description}">
-        <div class="alias-name">${name}</div>
-        <div class="alias-command-count">${commandCount} <span data-i18n="commands">commands</span></div>
-      </div>
-    `
-  }
-
-  selectAlias(aliasName) {
-    // Reuse the selectedKey property for alias selection
-    this.selectedKey = aliasName
-    this.renderAliasGrid()
-    this.renderCommandChain()
-    this.updateChainActions()
-  }
-
-  showAliasCreationModal() {
-    // Show a simplified modal for creating a new alias
-    const modal = this.createAliasCreationModal()
-    document.body.appendChild(modal)
-    modalManager.show('aliasCreationModal')
-  }
-
-  createAliasCreationModal() {
-    const modal = document.createElement('div')
-    modal.id = 'aliasCreationModal'
-    modal.className = 'modal'
-    modal.innerHTML = `
-      <div class="modal-content">
-        <div class="modal-header">
-          <h2 data-i18n="create_new_alias">Create New Alias</h2>
-          <button class="modal-close" data-modal="aliasCreationModal">
-            <i class="fas fa-times"></i>
-          </button>
-        </div>
-        <div class="modal-body">
-          <div class="form-group">
-            <label for="newAliasName" data-i18n="alias_name">Alias Name:</label>
-            <input type="text" id="newAliasName" class="form-control" placeholder="MyAlias" />
-          </div>
-          <div class="form-group">
-            <label for="newAliasDescription" data-i18n="description">Description:</label>
-            <input type="text" id="newAliasDescription" class="form-control" placeholder="Brief description" />
-          </div>
-        </div>
-        <div class="modal-footer">
-          <button class="btn btn-secondary" data-modal="aliasCreationModal" data-i18n="cancel">Cancel</button>
-          <button class="btn btn-primary" id="confirmCreateAliasBtn" data-i18n="create">Create</button>
-        </div>
-      </div>
-    `
-
-    // Add event listener for create button
-    modal.querySelector('#confirmCreateAliasBtn').addEventListener('click', () => {
-      const name = modal.querySelector('#newAliasName').value.trim()
-      const description = modal.querySelector('#newAliasDescription').value.trim()
-      
-      if (name) {
-        this.createAliasChain(name, description)
-        modalManager.hide('aliasCreationModal')
-        document.body.removeChild(modal)
-      }
-    })
-
-    return modal
-  }
-
-  createAliasChain(name, description = '') {
-    const profile = stoStorage.getProfile(this.currentProfile)
-    if (!profile) return
-
-    // Initialize aliases object if it doesn't exist
-    if (!profile.aliases) {
-      profile.aliases = {}
-    }
-
-    // Check if alias already exists
-    if (profile.aliases[name]) {
-      stoUI.showToast(i18next.t('alias_already_exists', {name: name}), 'error')
-      return
-    }
-
-    // Create new alias
-    profile.aliases[name] = {
-      description: description,
-      commands: ''
-    }
-
-    // Save profile
-    stoStorage.saveProfile(this.currentProfile, profile)
-    
-    // Update UI
-    this.renderAliasGrid()
-    this.selectAlias(name)
-    this.setModified(true)
-    
-    stoUI.showToast(i18next.t('alias_created', {name: name}), 'success')
-  }
-
-  async confirmDeleteAlias(aliasName) {
-    const confirmed = await stoUI.confirm(
-      i18next.t('confirm_delete_alias', { aliasName }),
-      i18next.t('delete_alias'),
-      'danger'
-    )
-
-    if (confirmed) {
-      this.deleteAliasChain(aliasName)
-    }
-  }
-
-  deleteAliasChain(aliasName) {
-    const profile = stoStorage.getProfile(this.currentProfile)
-    if (!profile || !profile.aliases || !profile.aliases[aliasName]) return
-
-    delete profile.aliases[aliasName]
-    stoStorage.saveProfile(this.currentProfile, profile)
-
-    // Clear selection if we deleted the selected alias
-    if (this.selectedKey === aliasName) {
-      this.selectedKey = null
-    }
-
-    this.renderAliasGrid()
-    this.renderCommandChain()
-    this.updateChainActions()
-    this.setModified(true)
-
-    stoUI.showToast(i18next.t('alias_deleted', {aliasName: aliasName}), 'success')
-  }
-
-  duplicateAlias(aliasName) {
-    const profile = stoStorage.getProfile(this.currentProfile)
-    if (!profile || !profile.aliases || !profile.aliases[aliasName]) return
-
-    const originalAlias = profile.aliases[aliasName]
-    
-    // Find a suitable new alias name
-    let newAliasName = aliasName + '_copy'
-    let counter = 1
-    
-    while (profile.aliases[newAliasName]) {
-      newAliasName = `${aliasName}_copy${counter}`
-      counter++
-    }
-
-    // Create duplicate
-    profile.aliases[newAliasName] = {
-      description: originalAlias.description + ' (copy)',
-      commands: originalAlias.commands
-    }
-
-    stoStorage.saveProfile(this.currentProfile, profile)
-    
-    this.renderAliasGrid()
-    this.selectAlias(newAliasName)
-    this.setModified(true)
-
-    stoUI.showToast(i18next.t('alias_created_from_template', {newAliasName: newAliasName}), 'success')
-  }
-
-  updateModeUI() {
-    if (this.currentEnvironment === 'alias') {
-      // Show alias view, hide key view
-      this.showAliasView()
-      this.renderAliasGrid()
-      this.renderCommandChain()
-      this.updateChainOptionsForAlias()
-    } else {
-      // Show key view, hide alias view
-      this.showKeyView()
-      this.renderKeyGrid()
-      this.renderCommandChain()
-      this.updateChainOptionsForKeybind()
-      this.filterCommandLibrary() // Apply environment filter to command library
+  updateToggleButtonVisibility() {
+    const toggleBtn = document.getElementById('toggleKeyViewBtn')
+    if (toggleBtn) {
+      // Hide toggle button in alias mode, show in keybind modes
+      toggleBtn.style.display = this.currentEnvironment === 'alias' ? 'none' : 'block'
     }
   }
 
@@ -2543,6 +1164,8 @@ export default class STOToolsKeybindManager {
       return
     }
 
+    // Simple grid view for aliases - no view modes needed
+    grid.classList.remove('categorized')
     grid.innerHTML = aliases.map(([name, alias]) => 
       this.createAliasChainElement(name, alias)
     ).join('')
@@ -5081,3 +3704,4 @@ export default class STOToolsKeybindManager {
 }
 
 // Initialize application
+Object.assign(STOToolsKeybindManager.prototype, profileManagement, keyHandling, uiRendering);
