+++ src/styles.css
@@ -1088,6 +1088,7 @@ body {
   padding: 0.5rem;
   border-radius: var(--radius);
   transition: var(--transition);
+  position: relative; /* Ensure proper positioning */
 }
 
 .category h4:hover {
@@ -1099,12 +1100,13 @@ body {
   width: 1.25rem;
 }
 
-/* Command Library category collapsible enhancements */
+/* Category collapsible enhancements */
 .category h4 .category-chevron {
   transition: transform 0.2s ease;
   color: var(--text-muted);
   font-size: 0.8rem;
   width: 12px;
+  flex-shrink: 0; /* Prevent chevron from shrinking */
 }
 
 .category h4:not(.collapsed) .category-chevron {
@@ -1115,23 +1117,53 @@ body {
   margin-bottom: 0;
 }
 
+.category h4 .key-count,
 .category h4 .command-count {
   margin-left: auto;
   font-size: 0.8rem;
   color: var(--text-muted);
   font-weight: normal;
+  flex-shrink: 0; /* Prevent count from shrinking */
 }
 
 .category-commands {
   display: grid;
   gap: 0.25rem;
   margin-left: 1.75rem;
+  overflow: hidden; /* Ensure proper containment */
 }
 
 .category-commands.collapsed {
   display: none;
 }
 
+/* Key grid categorized view specific styling */
+.key-grid.categorized {
+  display: block;
+  grid-template-columns: unset;
+  gap: unset;
+  overflow-y: auto;
+  overflow-x: hidden;
+}
+
+.key-grid.categorized .category {
+  margin-bottom: 1rem;
+}
+
+.key-grid.categorized .category:last-child {
+  margin-bottom: 0;
+}
+
+/* Alias grid specific styling */
+.alias-grid .category {
+  margin-bottom: 1rem;
+}
+
+.alias-grid .category:last-child {
+  margin-bottom: 0;
+}
+
+/* Command items within categories */
 .command-item {
   padding: 0.625rem 0.75rem;
   background: var(--surface-dark);
@@ -1172,6 +1204,41 @@ body {
 .command-item.vertigo-alias-item:hover {
   background: linear-gradient(135deg, #7c3aed, #5b21b6);
   transform: translateX(4px) scale(1.02);
+}
+
+/* Key-specific styling for command items */
+.command-item[data-key] .key-label {
+  font-weight: 500;
+  flex-shrink: 0;
+  transition: font-size 0.2s ease;
+}
+
+/* Dynamic font sizing for categorized view keys */
+.command-item[data-key][data-length='short'] .key-label {
+  font-size: 0.875rem;
+}
+
+.command-item[data-key][data-length='medium'] .key-label {
+  font-size: 0.8rem;
+}
+
+.command-item[data-key][data-length='long'] .key-label {
+  font-size: 0.75rem;
+}
+
+.command-item[data-key] .command-count-badge {
+  background: var(--accent);
+  color: white;
+  font-size: 0.6rem;
+  font-weight: 600;
+  padding: 0.125rem 0.375rem;
+  border-radius: 9999px;
+  min-width: 16px;
+  height: 16px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-left: auto;
 }
 
 /* Empty states */
@@ -4644,6 +4711,7 @@ input:checked + .slider:before {
   padding: 0.5rem;
   border-radius: var(--radius);
   transition: var(--transition);
+  position: relative; /* Ensure proper positioning */
 }
 
 .category h4:hover {
@@ -4655,12 +4723,13 @@ input:checked + .slider:before {
   width: 1.25rem;
 }
 
-/* Command Library category collapsible enhancements */
+/* Category collapsible enhancements */
 .category h4 .category-chevron {
   transition: transform 0.2s ease;
   color: var(--text-muted);
   font-size: 0.8rem;
   width: 12px;
+  flex-shrink: 0; /* Prevent chevron from shrinking */
 }
 
 .category h4:not(.collapsed) .category-chevron {
@@ -4671,23 +4740,53 @@ input:checked + .slider:before {
   margin-bottom: 0;
 }
 
+.category h4 .key-count,
 .category h4 .command-count {
   margin-left: auto;
   font-size: 0.8rem;
   color: var(--text-muted);
   font-weight: normal;
+  flex-shrink: 0; /* Prevent count from shrinking */
 }
 
 .category-commands {
   display: grid;
   gap: 0.25rem;
   margin-left: 1.75rem;
+  overflow: hidden; /* Ensure proper containment */
 }
 
 .category-commands.collapsed {
   display: none;
 }
 
+/* Key grid categorized view specific styling */
+.key-grid.categorized {
+  display: block;
+  grid-template-columns: unset;
+  gap: unset;
+  overflow-y: auto;
+  overflow-x: hidden;
+}
+
+.key-grid.categorized .category {
+  margin-bottom: 1rem;
+}
+
+.key-grid.categorized .category:last-child {
+  margin-bottom: 0;
+}
+
+/* Alias grid specific styling */
+.alias-grid .category {
+  margin-bottom: 1rem;
+}
+
+.alias-grid .category:last-child {
+  margin-bottom: 0;
+}
+
+/* Command items within categories */
 .command-item {
   padding: 0.625rem 0.75rem;
   background: var(--surface-dark);
@@ -4728,6 +4827,41 @@ input:checked + .slider:before {
 .command-item.vertigo-alias-item:hover {
   background: linear-gradient(135deg, #7c3aed, #5b21b6);
   transform: translateX(4px) scale(1.02);
+}
+
+/* Key-specific styling for command items */
+.command-item[data-key] .key-label {
+  font-weight: 500;
+  flex-shrink: 0;
+  transition: font-size 0.2s ease;
+}
+
+/* Dynamic font sizing for categorized view keys */
+.command-item[data-key][data-length='short'] .key-label {
+  font-size: 0.875rem;
+}
+
+.command-item[data-key][data-length='medium'] .key-label {
+  font-size: 0.8rem;
+}
+
+.command-item[data-key][data-length='long'] .key-label {
+  font-size: 0.75rem;
+}
+
+.command-item[data-key] .command-count-badge {
+  background: var(--accent);
+  color: white;
+  font-size: 0.6rem;
+  font-weight: 600;
+  padding: 0.125rem 0.375rem;
+  border-radius: 9999px;
+  min-width: 16px;
+  height: 16px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-left: auto;
 }
 
 /* Empty states */
